==> 05 Pairing Folders & Page_ids With Row Indices  & Categories.ipynb <==
### This file:
<br>
- Opens text files containing page ids and categories
- Creates dictionary "d" where keys are page ids, values are major categories ("nagykategoriak")
- Opens list of pages with likes (the 'not_empty_folders')
- Creates the dictionary "page_categories" where these not empty folders are paired with thier respectivy categories
- Saves the dict into "page_categories.pkl"

import csv
import pickle

kat_files = ["1csaladotthon.txt", "2egeszsegsport.txt", "3elmenyszorakozas.txt", "4kommersz.txt",
            "5kozelet.txt", "6technika.txt", "7tudatossag.txt", "8uzletielet.txt"]
kategoriak = [1, 2, 3, 4, 5, 6, 7, 8]

d = {}
for i in range(0,8):
    f = open(kat_files[i], 'r')
    l = f.readlines()
    f.close()
    for j in range(0, len(l)):
        h = len(l[j])
        page = int(l[j][2:h-1])
        d[page] = kategoriak[i]


len(d)

page_ids = []
with open("page_ids.csv", 'r') as csvfile:
    input = csv.reader(csvfile, delimiter = ",")
    for row in input:
        page_ids.append(row)
csvfile.close()
    

page_categories = {} # a dict for pages with likes where keys are page_ids and values are categories (nagykategoriak)
for i in range(0, len(page_ids)):
    key = page_ids[i][0]
    page_categories[key] = d[int(key)]

len(page_categories)

f = open("page_categories.pkl","wb")
pickle.dump(page_categories,f)
f.close()
