==> 16 Scipy Hierarchical Cluster.ipynb <==
from scipy.sparse import coo_matrix
from scipy.sparse import save_npz, load_npz
%matplotlib inline
import matplotlib.pyplot as plt
import numpy as np
import pandas as pd
from sklearn.cluster import AgglomerativeClustering
from sklearn.cluster import spectral_clustering
from scipy.cluster.hierarchy import linkage, dendrogram, fcluster
import pickle

<br>Import matrices: *mx_upper* as upper triangle matrix, *mx* as symmetirc matrix. 

mx_upper = load_npz('C:\\FOLDERS\\CEU\\Capstone Project\\Data\\sparse_matrix_upper_triangle.npz')

mx = load_npz('C:\\FOLDERS\\CEU\\Capstone Project\\Data\\sparse_matrix.npz')

mx_array = mx.toarray()

from scipy.spatial.distance import squareform
mx_s = squareform(mx)

g = open("page_categories.pkl", "rb")
temp = pickle.load(g)
g.close()

categories = []

for key in temp:
    categories.append(temp[key])

<br>
### Scipy Hierarchy Clustering

Getting upper triangular values.

l = [] # list for upper triangular (ex-diagonal) values

for i in range(0, 8743):
    for j in range(i + 1, 8743):
        l.append(mx_array[i,j])

Z = linkage(l, method="ward")

k=8
scipy_hier_labels = fcluster(Z, k, criterion='maxclust')

page_labels = pd.DataFrame(scipy_hier_labels, columns=["labels"])
cat = pd.Series(categories)
page_labels['category'] = cat.values

page_labels.head()

ct = pd.crosstab(page_labels["labels"], page_labels["category"])
ct
