==> 31 Dissimilarity Matrices Pages Ex Cat 7.ipynb <==
from scipy.sparse import coo_matrix
from scipy.sparse import save_npz, load_npz
%matplotlib inline
import matplotlib
import matplotlib.pyplot as plt
import numpy as np
import pandas as pd
from scipy.cluster.hierarchy import linkage, dendrogram, fcluster
import pickle
import time
from scipy import spatial

<span> 1. </span> Import sparse symmetric matrix.

mx = load_npz('C:\\FOLDERS\\CEU\\Capstone Project\\Data\\active_users_sparse_matrix_ex_7.npz')

mx_array = mx.toarray()

<span>2.</span> Retreive categories data, omit cat 7, and sort pages in alphabetical order. 

g = open("page_categories.pkl", "rb")
page_categories = pickle.load(g)
g.close()

f = open("C:\\FOLDERS\\CEU\\Capstone Project\\Data\\active_users_page_likes.pkl", "rb")
temp = pickle.load(f)
f.close()

active_page_ids = []

for key in temp.keys():
    active_page_ids.append(key)

active_page_ids = sorted(active_page_ids)

active_page_categories = {}

for key in page_categories.keys():
    if key in active_page_ids:
        active_page_categories[key] = page_categories[key]
        
active_page_categories_ex_7 = {}

for key in active_page_categories.keys():
    if active_page_categories[key] != 7:
        active_page_categories_ex_7[key] = active_page_categories[key]

<span>3.</span> Save page categories dict as pickle for later use.

f = open("C:\\FOLDERS\\CEU\\Capstone Project\\Data\\active_page_categories_ex_7.pkl", "wb")
pickle.dump(active_page_categories_ex_7,f)
f.close()

f = open("C:\\FOLDERS\\CEU\\Capstone Project\\Data\\active_page_categories_ex_7.pkl", "rb")
active_page_categories_ex_7 = pickle.load(f)
f.close()

categories = []

for key in active_page_categories_ex_7:
    categories.append(active_page_categories_ex_7[key])

len(categories)

mx_corr = np.zeros((5911, 5911))
mx_diss = np.zeros((5911, 5911))
mx_cos =  np.zeros((5911, 5911))

t1 = time.time()

for i in range(0,5910):
    for j in range(i+1, 5911):
        mx_diss[i,j] = 100 - 100 * int(mx_array[i,j]) * int(mx_array[i,j]) / (int(np.max(mx_array[i,:])) * int(np.max(mx_array[:,j])))

print(round(time.time()-t1) / 60, 1)


t1 = time.time()

for i in range(0,5910):
    for j in range(i+1, 5911):
        mx_cos[i,j] = 100 * spatial.distance.cosine(mx_array[i,:], mx_array[:,j])

print(round(time.time()-t1) / 60, 1)

t1 = time.time()

for i in range(0, 5910):
    for j in range(i+1, 5911):
        mx_corr[i,j] = 1 - np.corrcoef(mx_array[i, :], mx_array[: , j])[1,0]

print(round(time.time()-t1) / 60, 1)

np.savetxt("C:\\FOLDERS\\CEU\\Capstone Project\\Data\\active_users_ex_7_mx_correlation_dissimilarity.gz",
           mx_corr, delimiter=',')

np.savetxt("C:\\FOLDERS\\CEU\\Capstone Project\\Data\\active_users_ex_7_mx_cosine_dissimilarity.gz",
           mx_cos, delimiter=',')

np.savetxt("C:\\FOLDERS\\CEU\\Capstone Project\\Data\\active_users_ex_7_mx_square_dissimilarity.gz",
           mx_diss, delimiter=',')
